Refactor views into seperate files
