+++ moodmusic/dashboard/views/__init__.py
@@ -1,0 +1,3 @@
+from .static import *  # noqa
+from .data_drop import *  # noqa
+from .twilio_verify import *  # noqa
